moved coherence and topic diff to utility
